+++ worker/src/handlers/statement/statementForwarding/tests/statementForwardingRequestHandler-test.js
@@ -1,12 +1,21 @@
 import StatementForwarding from 'lib/models/statementForwarding';
 import Statement from 'lib/models/statement';
+import Promise from 'bluebird';
 import nock from 'nock';
-import async from 'async';
 import { expect } from 'chai';
+// import ForwardingRequestError from
+//   'worker/handlers/statement/statementForwarding/ForwardingRequestError';
 import statementForwardingRequestHandler from '../statementForwardingRequestHandler';
 
+const promiseRequestHandler = Promise.promisify(statementForwardingRequestHandler);
+
 describe('Statement Forwarding Request', () => {
-  it('Request success', (done) => {
+  afterEach(async () => {
+    await Statement.remove({});
+    await StatementForwarding.remove({});
+  });
+
+  it('Request returns 308', async () => {
     const statementId = '59438cabedcedb70146337ec';
     const statementForwardingId = '59438cabedcedb70146337eb';
     const organisationId = '561a679c0c5d017e4004715a';
@@ -17,8 +26,8 @@ describe('Statement Forwarding Request', () => {
       organisation: organisationId,
       active: true,
       configuration: {
-        protocal: 'http',
-        url: 'localhost:3101/',
+        protocol: 'http',
+        url: 'redirectto/',
         method: 'POST'
       }
     };
@@ -31,68 +40,36 @@ describe('Statement Forwarding Request', () => {
       }
     };
 
-    Statement.create(statement)
-    .then(() => {
-      const request = nock('http://localhost:3101/')
-        .post('/', {
-          test: 'test'
-        })
-        .reply(200, {
-          _id: '1',
-          _rev: '1',
-          success: true
-        });
-      return { request };
-    }).then((params) => {
-      const promise = new Promise(resolve =>
-        statementForwardingRequestHandler({
-          statementForwarding,
-          statement
-        }, () => {
-          resolve(params);
-        })
-      );
+    const request = nock('http://localhost:3101/')
+      .post('/', {
+        test: 'test'
+      }).reply(200, {
+        _id: '1',
+        _rev: '1',
+        success: true
+      });
 
-      return promise;
-    })
-    .then(({ request }) => {
-      expect(request.isDone()).to.equal(true);
-    })
-    .then(() =>
-      new Promise((resolve) => {
-        Statement.findById(statementId).then((statementModel) => {
-          expect(statementModel.pendingForwardingQueue.length).to
-            .equal(0);
-          expect(statementModel.completedForwardingQueue[0].toString()).to
-            .equal(statementForwardingId);
-          resolve();
-        });
-      })
-    )
-    .then(() => new Promise(reslove =>
-      async.forEach(
-        [StatementForwarding, Statement],
-        (model, doneDeleting) => {
-          model.remove({}, doneDeleting);
-        },
-        reslove
-      )
-    ), () => new Promise(reslove =>
-      async.forEach(
-        [StatementForwarding, Statement],
-        (model, doneDeleting) => {
-          model.remove({}, doneDeleting);
-        },
-        reslove
-      )
-    ))
-    .then(() => {
-      done();
-    });
+    const redirectRequest = nock('http://redirectto/')
+      .post('/', {
+        test: 'test'
+      }).reply(308, undefined, {
+        Location: 'http://localhost:3101/'
+      });
+
+    await Statement.create(statement);
+
+    await promiseRequestHandler({ statementForwarding, statement });
+    Promise.delay(1000);
+    expect(redirectRequest.isDone()).to.equal(true);
+    expect(request.isDone()).to.equal(true);
+
+    const doneStatement = await Statement.findById(statementId);
+    expect(doneStatement.pendingForwardingQueue.length).to.equal(0);
+    expect(doneStatement.completedForwardingQueue.length).to.equal(1);
+    expect(doneStatement.completedForwardingQueue[0].toString()).to.equal(statementForwardingId);
   }).timeout(10000);
 
-
-  it('If a request fails, keep in pending', (done) => {
+  it('Request success', async () => {
     const statementId = '59438cabedcedb70146337ec';
     const statementForwardingId = '59438cabedcedb70146337eb';
     const organisationId = '561a679c0c5d017e4004715a';
@@ -117,57 +94,77 @@ describe('Statement Forwarding Request', () => {
       }
     };
 
-    Statement.create(statement)
-    .then(() => {
-      const request = nock('http://localhost:3101')
-        .post('/', {
-          test: 'test'
-        })
-        .reply(404, {
-          _id: '1',
-          _rev: '1',
-          success: false
-        });
-      return { request };
-    }).then((params) => {
-      const promise = new Promise(resolve =>
-        statementForwardingRequestHandler({
-          statementForwarding,
-          statement
-        }, () => {
-          resolve(params);
-        })
-      );
-
-      return promise;
-    })
-    .then(({ request }) => {
-      expect(request.isDone()).to.equal(true);
-    })
-    .then(() =>
-      new Promise((resolve) => {
-        Statement.findById(statementId).then((statementModel) => {
-          expect(statementModel.pendingForwardingQueue[0].toString()).to
-            .equal(statementForwardingId);
-          expect(statementModel.completedForwardingQueue.length).to
-            .equal(0);
-          expect(statementModel.failedForwardingLog[0].statementForwarding_id.toString()).to
-            .equal(statementForwardingId);
-          resolve();
-        });
+    const request = nock('http://localhost:3101/')
+      .post('/', {
+        test: 'test'
       })
-    )
-    .then(() => new Promise(reslove =>
-      async.forEach(
-        [StatementForwarding, Statement],
-        (model, doneDeleting) => {
-          model.remove({}, doneDeleting);
-        },
-        reslove
-      )
-    ))
-    .then(() => {
-      done();
-    });
+      .reply(200, {
+        _id: '1',
+        _rev: '1',
+        success: true
+      });
+
+    await Statement.create(statement);
+
+    await promiseRequestHandler({ statementForwarding, statement });
+    expect(request.isDone()).to.equal(true);
+
+    const doneStatement = await Statement.findById(statementId);
+    expect(doneStatement.pendingForwardingQueue.length).to.equal(0);
+    expect(doneStatement.completedForwardingQueue.length).to.equal(1);
+    expect(doneStatement.completedForwardingQueue[0].toString()).to.equal(statementForwardingId);
+  }).timeout(10000);
+
+
+  it('If a request fails, keep in pending', async () => {
+    const statementId = '59438cabedcedb70146337ec';
+    const statementForwardingId = '59438cabedcedb70146337eb';
+    const organisationId = '561a679c0c5d017e4004715a';
+
+    const statementForwarding = {
+      _id: statementForwardingId,
+      lrs_id: '560a679c0c5d017e4004714f',
+      organisation: organisationId,
+      active: true,
+      configuration: {
+        protocol: 'http',
+        url: 'localhost:3101/',
+        method: 'POST'
+      }
+    };
+
+    const statement = {
+      _id: statementId,
+      organisation: organisationId,
+      statement: {
+        test: 'test'
+      }
+    };
+
+    await Statement.create(statement);
+
+    const request = nock('http://localhost:3101')
+      .post('/', {
+        test: 'test'
+      })
+      .reply(404, {
+        _id: '1',
+        _rev: '1',
+        success: false
+      });
+    try {
+      await promiseRequestHandler({ statementForwarding, statement });
+    } catch (err) {
+      expect(err).to.be.err; // eslint-disable-line no-unused-expressions
+    }
+    expect(request.isDone()).to.equal(true);
+
+    const doneStatement = await Statement.findById(statementId);
+    expect(doneStatement.pendingForwardingQueue[0].toString()).to
+      .equal(statementForwardingId);
+    expect(doneStatement.completedForwardingQueue.length).to
+      .equal(0);
+    expect(doneStatement.failedForwardingLog[0].statementForwarding_id.toString()).to
+      .equal(statementForwardingId);
   }).timeout(5000);
 });