+++ ui/src/utils/createAsyncDuck.js
@@ -2,6 +2,8 @@ import { put, call, takeEvery } from 'redux-saga/effects';
 import Promise from 'bluebird';
 import identity from 'lodash/identity';
 import DispatchNotReadyError from 'ui/utils/errors/DispatchNotReadyError';
+import Unauthorised from 'lib/errors/Unauthorised';
+import { actions as logoutAcitons } from 'ui/redux/modules/auth/logout';
 import { alert } from 'ui/redux/modules/alerts';
 
 const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
@@ -90,6 +92,9 @@ export default function createAsyncDuck({
       if (failureDelay > -1) {
         yield call(delay, failureDelay); // shows the failure state for 2 seconds
         yield put(actions.complete(args));
+      }
+      if (err.constructor === Unauthorised) {
+        yield put(logoutAcitons.logoutAction());
       }
     }
   }