+++ ui/src/containers/SideNav/CollapsibleNav.js
@@ -43,6 +43,12 @@ class CollapsibleNav extends Component {
     onClick(routeName);
   }
 
+  animationComplete = isExpanded => () => {
+    if (isExpanded) {
+      this.bodyDom.style.height = 'auto';
+    }
+  }
+
   render = () => {
     const { bodyHeight } = this.state;
     const {
@@ -73,7 +79,10 @@ class CollapsibleNav extends Component {
           <i className={icon} /> {label}
         </Link>
 
-        <VelocityComponent animation={animationProps} duration={350}>
+        <VelocityComponent
+          animation={animationProps}
+          duration={350}
+          complete={this.animationComplete(isExpanded)} >
           <ul
             className={`nav animated fadeIn ${styles.subNav}`}
             ref={this.saveBodyRef}