+++ ui/src/redux/modules/pagination/fetchModels.js
@@ -10,6 +10,7 @@ import { normalize, arrayOf } from 'normalizr';
 import entityReviver from 'ui/redux/modules/models/entityReviver';
 import * as mergeEntitiesDuck from 'ui/redux/modules/models/mergeEntities';
 import { IN_PROGRESS, COMPLETED, FAILED } from 'ui/utils/constants';
+import Unauthorised from 'lib/errors/Unauthorised';
 import diffEdges from './fetchModelsDiff';
 
 
@@ -24,6 +25,7 @@ export const BACKWARD = 'BACKWARD'; // backward pagination direction
  * SELECTORS
  */
 const cacheDuration = moment.duration({ minute: 3 });
+const failedCacheDuration = moment.duration({ minute: 1 });
 
 const defaultFilter = new Map();
 const defaultSort = new Map({ createdAt: -1, _id: 1 });
@@ -76,6 +78,14 @@ const shouldFetchSelector = ({ schema, filter, sort, cursor }) => (createSelecto
     if (requestState === IN_PROGRESS) {
       return false;
     }
+
+    if (
+      requestState === FAILED &&
+      moment().diff(cachedAt) < failedCacheDuration.asMilliseconds()
+    ) {
+      return false;
+    }
+
     const cachedFor = moment().diff(cachedAt);
     if (cachedFor < cacheDuration.asMilliseconds()) {
       return false;
@@ -169,14 +179,15 @@ export const reduceSuccess = (
 
 const fetchModels = createAsyncDuck({
   actionName: 'learninglocker/pagination/FETCH_MODELS',
-  failureDelay: 20000,
+  failureDelay: 2000,
 
   reduceStart,
 
   reduceSuccess,
 
   reduceFailure: (state, { schema, filter, sort, cursor }) =>
-    state.setIn([schema, filter, sort, cursor, 'requestState'], FAILED),
+    state.setIn([schema, filter, sort, cursor, 'requestState'], FAILED)
+      .setIn([schema, filter, sort, cursor, 'cachedAt'], moment()),
 
   reduceComplete: (state, { schema, filter, sort, cursor }) =>
     state.setIn([schema, filter, sort, cursor, 'requestState'], null),
@@ -252,7 +263,13 @@ const fetchModels = createAsyncDuck({
       first,
       last
     });
-    if (status >= 300) throw new Error(body.message || body);
+
+    if (status === 401) {
+      throw new Unauthorised('Unauthorised');
+    }
+    if (status >= 300) {
+      throw new Error(body.message || body);
+    }
 
     const models = map(body.edges, 'node');
     const ids = map(models, '_id');