+++ ui/src/utils/hocs/withModels.js
@@ -17,13 +17,19 @@ import {
 
 export default (WrappedComponent) => {
   class WithModels extends Component {
-    componentWillMount = () => this.fetchModels(this.props)
-    componentWillReceiveProps = nextProps => this.fetchModels(nextProps)
+    componentWillMount = () => {
+      return this.fetchModels(this.props);
+    }
+
+    componentWillReceiveProps = nextProps =>
+      this.fetchModels(nextProps);
 
     fetchModels = ({ schema, filter, sort, first }) => {
       if (filter) {
         this.props.fetchAllOutstandingModels({ schema, filter, sort, first })
-          .catch(err => console.error(err));
+          .catch((err) => {
+            console.error(err);
+          });
       }
     }
     addModel = (args) => {
@@ -42,14 +48,14 @@ export default (WrappedComponent) => {
       const { schema, filter, sort, first } = this.props;
       return this.props.fetchMore({ schema, filter, sort, first, ...args });
     }
-    render = () => (
-      <WrappedComponent
+    render = () =>
+      (<WrappedComponent
         {...this.props}
         fetchMore={this.fetchMore}
         addModel={this.addModel}
         deleteModel={this.deleteModel}
         updateModel={this.updateModel} />
-    )
+      )
   }
   return connect((state, { schema, filter, sort, cursor }) => (
     {