+++ CHANGELOG.md
@@ -9,7 +9,8 @@ Format based on [Keep a Changelog](http://keepachangelog.com/)
 ### Deprecated
 ### Removed
 ### Fixed
- - 'statement forwarding' menu item displaying incorrectly
+ - Allowed stamenent forwarding to follow redirects 
+ - Statement forwarding menu item displaying incorrectly
  - Client auth not working with some APIs
 ### Security
 ### Migrations